Combine MLA index-cache layout and builder tests.
Keep CPU layout cases runnable without AITER, and skip builder/ModelRunner
cases when the MLA backend cannot import.
